pyproject.toml
[project]
name = "multi-model-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.86.0",
    "ipython>=9.12.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
